I added my function(s) to the file.
I will need to talk to the person who implemented to_binary since my thing heavily relies on that. I tested it for bugs and all and am confident it works with all inputs. If there happen to be any errors please contact me (Kyle N.) so I can address them and fix them. I am confident that any and all bugs that arise would occur because of a difference in how the to_binary function is made (wether the first index is the least significant bit or the most significant bit), so if a problem like that arises, I can jut change the indexes and the outputs will work again.